chore: refactor helper functions to not be tied to completions FUI-1226
