+++ src/plugin/completions/helpers.test.ts
@@ -5,145 +5,8 @@ import {
   constructGlobalAriaCompletion,
   constructGlobalAttrCompletion,
   constructGlobalEventCompletion,
-  getCompletionType,
 } from './helpers';
-
-describe('getCompletionType', () => {
-  const tests: [string, [TemplateContext, LineAndCharacter], any][] = [
-    [
-      'Key "none" for a blank template',
-      [html``, { line: 0, character: 0 }],
-      { key: 'none', params: undefined },
-    ],
-    [
-      'Key "none" after a fully closed custom element',
-      [
-        html`
-          <c-e></c-e>
-        `,
-        { line: 0, character: 11 },
-      ],
-      { key: 'none', params: undefined },
-    ],
-    [
-      'Key "none" after a fully closed element',
-      [
-        html`
-          <div></div>
-        `,
-        { line: 0, character: 11 },
-      ],
-      { key: 'none', params: undefined },
-    ],
-    [
-      'Key "none" after a self closing element',
-      [
-        html`
-          <input />
-        `,
-        { line: 0, character: 9 },
-      ],
-      { key: 'none', params: undefined },
-    ],
-    [
-      'Key "none" after in invalid empty tag',
-      [
-        html`
-          <>
-        `,
-        { line: 0, character: 2 },
-      ],
-      { key: 'none', params: undefined },
-    ],
-    [
-      'Key "none" after an unclosed tag',
-      [
-        html`
-          <div></div>
-        `,
-        { line: 0, character: 5 },
-      ],
-      { key: 'none', params: undefined },
-    ],
-    // custom-element-name
-    [
-      'Key "custom-element-name" after an opening tag',
-      [
-        html`
-          <
-        `,
-        { line: 1, character: 11 },
-      ],
-      { key: 'custom-element-name', params: undefined },
-    ],
-    [
-      'Key "custom-element-name" after an opening with characters',
-      [html`<di`, { line: 0, character: 3 }],
-      { key: 'custom-element-name', params: undefined },
-    ],
-    [
-      'Key "custom-element-name" after an opening with characters, after a valid tag',
-      [html`<div></div><di`, { line: 0, character: 14 }],
-      { key: 'custom-element-name', params: undefined },
-    ],
-    // custom-element-attribute, with custom element name
-    [
-      'Key "custom-element-attribute" after the start of a custom element, and the name',
-      [html`<cus-`, { line: 0, character: 5 }],
-      { key: 'custom-element-attribute', params: 'cus-' },
-    ],
-    [
-      'Key "custom-element-attribute" after a custom element, and the name',
-      [html`<cus-elem`, { line: 0, character: 9 }],
-      { key: 'custom-element-attribute', params: 'cus-elem' },
-    ],
-    [
-      'Key "custom-element-attribute" after a custom element accounting for another custom element',
-      [html`<ce-elem></ce-elem><cus-elem`, { line: 0, character: 28 }],
-      { key: 'custom-element-attribute', params: 'cus-elem' },
-    ],
-    [
-      'Key "custom-element-attribute" and gets custom element name when attribute is present',
-      [html`<ce-elem></ce-elem><cus-elem attr`, { line: 0, character: 33 }],
-      { key: 'custom-element-attribute', params: 'cus-elem' },
-    ],
-    [
-      'Key "custom-element-attribute" and gets custom element name when string attribute is present',
-      [html`<ce-elem></ce-elem><cus-elem attr="value"`, { line: 0, character: 41 }],
-      { key: 'custom-element-attribute', params: 'cus-elem' },
-    ],
-    [
-      'Key "custom-element-attribute" and gets custom element when cursor is inside of a fully closed element',
-      [
-        html`
-          <ce-elem></ce-elem>
-          <cus-elem attr="value"></cus-elem>
-        `,
-        { line: 2, character: 32 },
-      ],
-      { key: 'custom-element-attribute', params: 'cus-elem' },
-    ],
-    [
-      'Key "custom-element-attribute" and gets custom element when cursor is on the next line to an opening tag',
-      [
-        html`
-          <cus-elem
-          `,
-        { line: 2, character: 2 },
-      ],
-      { key: 'custom-element-attribute', params: 'cus-elem' },
-    ],
-  ];
-
-  for (const test of tests) {
-    const [name, [context, lineAndChar], expected] = test;
-    it(name, () => {
-      // `getComptionType` is private so need to case to `any` to test it
-      const result = getCompletionType(context, lineAndChar);
-      expect(result).toEqual(expected);
-    });
-  }
-});
+import { getTokenTypeWithInfo } from '../utils';
 
 describe('constructGlobalAriaCompletion', () => {
   it('returns CompletionEntry', () => {
