+++ src/plugin/completions/helpers.ts
@@ -1,61 +1,5 @@
-// TODO: Unit tests!
-
-import {
-  CompletionEntry,
-  LineAndCharacter,
-  ScriptElementKind,
-} from 'typescript/lib/tsserverlibrary';
-import { TemplateContext } from 'typescript-template-language-service-decorator';
+import { CompletionEntry, ScriptElementKind } from 'typescript/lib/tsserverlibrary';
 import { GlobalAttrType } from '../global-data/global-data.types';
-import { replaceTemplateStringBinding } from '../utils';
-import { CompletionTypeParams } from './completions.types';
-
-// Unclosed tag, suggest any tag names
-export const suggestTags = (line: string) => unfinishedTag.test(line);
-
-// Returns the currently written custom element tag name, or false if not found
-export const suggestCustomElements = (line: string) => {
-  if (!unfinishedCustomElement.test(line)) return false;
-  const lastTag = line.split(/</).pop() as string;
-  return lastTag.split(' ')[0].trim();
-};
-
-const unfinishedTag = /<[^>]*$/;
-
-// Like an unfinished tag, but with a hyphen
-const unfinishedCustomElement = /<.+-[^>]*$/m;
-
-export function getCompletionType(
-  context: TemplateContext,
-  position: LineAndCharacter
-): CompletionTypeParams {
-  const rawLine = context.rawText.split(/\n/g)[position.line];
-  const processedLine = replaceTemplateStringBinding(
-    context.rawText.substring(0, context.toOffset(position))
-  );
-
-  {
-    let tagname: string | false;
-    if (((tagname = suggestCustomElements(processedLine)), tagname)) {
-      return {
-        key: 'custom-element-attribute',
-        params: tagname,
-      };
-    }
-  }
-
-  if (suggestTags(rawLine)) {
-    return {
-      key: 'custom-element-name',
-      params: undefined,
-    };
-  }
-
-  return {
-    key: 'none',
-    params: undefined,
-  };
-}
 
 export function constructGlobalAriaCompletion(name: string): CompletionEntry {
   return {
